+++ backend/app/api/routes/__init__.py
@@ -1,0 +1,1 @@
+# This file ensures Python treats the directory as a package
