+++ backend/app/api/routes/general.py
@@ -1,0 +1,31 @@
+from datetime import datetime
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+from pydantic import BaseModel
+
+from app.database import get_db
+from app.models import Appointment
+
+router = APIRouter(
+    prefix="",
+    tags=["general"],
+)
+# Request model for appointment creation
+class AppointmentCreate(BaseModel):
+    name: str
+    created_at: datetime
+
+@router.post("/appointments")
+def create_appointment(appointment: AppointmentCreate, db: Session = Depends(get_db)):
+    try:
+        db_appointment = Appointment(
+            name=appointment.name,
+            created_at=appointment.created_at,
+            is_deleted=False
+        )
+        db.add(db_appointment)
+        db.commit()
+        db.refresh(db_appointment)
+        return db_appointment
+    except Exception as e:
+        raise HTTPException(status_code=400, detail=str(e))
